Refactor CLI Acquisition Context
