+++ src/client/lib/CliAcquisition.ts
@@ -10,31 +10,17 @@ import * as fs from 'fs-extra';
 import * as glob from 'glob';
 import * as os from 'os';
 import { Extract } from 'unzip-stream'
-import { ITelemetry } from '../../common/OneDSLoggerTelemetry/telemetry/ITelemetry';
 import find from 'find-process';
 import { spawnSync } from 'child_process';
 import commandExists from 'command-exists';
 import { oneDSLoggerWrapper } from '../../common/OneDSLoggerTelemetry/oneDSLoggerWrapper';
-
-// allow for DI without direct reference to vscode's d.ts file: that definintions file is being generated at VS Code runtime
-export interface ICliAcquisitionContext {
-    readonly extensionPath: string;
-    readonly globalStorageLocalPath: string;
-    readonly telemetry: ITelemetry;
-    showInformationMessage(message: string, ...items: string[]): void;
-    showErrorMessage(message: string, ...items: string[]): void;
-    showCliPreparingMessage(version: string): void;
-    showCliReadyMessage(): void;
-    showCliInstallFailedError(err: string): void;
-    locDotnetNotInstalledOrInsufficient(): string;
-}
+import { ICliAcquisitionContext } from './CliAcquisitionContext';
 
 export interface IDisposable {
     dispose(): void;
 }
 
 export class CliAcquisition implements IDisposable {
-
     private readonly _context: ICliAcquisitionContext;
     private readonly _cliPath: string;
     private readonly _cliVersion: string;
@@ -241,4 +227,3 @@ export class CliAcquisition implements IDisposable {
         }
     }
 }
-
